# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateOutputRequest": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "CreateOutputRequest",
        "type": "object"
      },
...
      "EditTarget": {
        "description": "What ``/edit/{output_id}`` needs to reopen the customizer.",
        "properties": {
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "output_id": {
            "title": "Output Id",
            "type": "string"
          },
          "params": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ]
            },
            "title": "Params",
            "type": "object"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "source": {
            "enum": [
              "record",
              "3mf"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "output_id",
          "slug",
          "name",
          "source"
        ],
        "title": "EditTarget",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OutputDetail": {
        "properties": {
          "bbox_mm": {
            "$ref": "#/components/schemas/BoundingBox"
          },
          "colors": {
            "items": {
              "type": "string"
            },
            "title": "Colors",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "has_thumbnail": {
            "title": "Has Thumbnail",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "library_file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Library File Id"
          },
          "library_file_plate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Library File Plate"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "params": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ]
            },
            "title": "Params",
            "type": "object"
          },
          "parts": {
            "items": {
              "$ref": "#/components/schemas/PartInfo"
            },
            "title": "Parts",
            "type": "array"
          },
          "pipeline_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Run Id"
          },
          "plates": {
            "items": {
              "$ref": "#/components/schemas/PlateSend"
            },
            "title": "Plates",
            "type": "array"
          },
          "print_route": {
            "anyOf": [
              {
                "enum": [
                  "pipeline",
                  "slice_queue"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Print Route"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "queue_item_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Item Id"
          },
          "slice_job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slice Job Id"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "id",
          "slug",
          "job_id",
          "created_at",
          "bbox_mm",
          "has_thumbnail"
        ],
        "title": "OutputDetail",
        "type": "object"
      },
      "OutputPlate": {
        "description": "One plate of the output's 3MF (#83): the ``plate_id`` a print of it queues.",
        "properties": {
          "has_thumbnail": {
            "title": "Has Thumbnail",
            "type": "boolean"
          },
          "index": {
            "title": "Index",
            "type": "integer"
          }
        },
        "required": [
          "index",
          "has_thumbnail"
        ],
        "title": "OutputPlate",
        "type": "object"
      },
...
      "SendRequest": {
        "properties": {
          "copies": {
            "anyOf": [
              {
                "maximum": 1000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Copies"
          },
          "mode": {
            "default": "library",
            "enum": [
              "library",
              "queue"
            ],
            "title": "Mode",
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/PrintOptions"
          }
        },
        "title": "SendRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Self-hosted OpenSCAD customizer for Bambuddy.",
    "title": "ScadBuddy",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/models/{slug}/outputs": {
      "get": {
        "description": "Details, not summaries: the history page shows each output's parameter diff, and\na summary list would make it fetch every row again one at a time.",
        "operationId": "list_outputs_api_v1_models__slug__outputs_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "maxLength": 108,
              "pattern": "^(builtin:)?[a-z0-9][a-z0-9-]*$",
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OutputDetail"
                  },
                  "title": "Response List Outputs Api V1 Models  Slug  Outputs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Output history",
        "tags": [
          "outputs"
        ]
      },
      "post": {
        "operationId": "create_output_api_v1_models__slug__outputs_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "maxLength": 108,
              "pattern": "^(builtin:)?[a-z0-9][a-z0-9-]*$",
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOutputRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutputDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Persist a finished render",
        "tags": [
          "outputs"
        ]
      }
    },
...
    "/api/v1/outputs/{output_id}": {
      "delete": {
        "operationId": "delete_output_api_v1_outputs__output_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "output_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Output Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete an output",
        "tags": [
          "outputs"
        ]
      },
      "get": {
        "operationId": "get_output_api_v1_outputs__output_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "output_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Output Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutputDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Output detail",
        "tags": [
          "outputs"
        ]
      }
    },
    "/api/v1/outputs/{output_id}/edit": {
      "get": {
        "description": "Where ``/edit/{output_id}`` should land, and with which values.\n\nThe record answers first. When it is gone \u2014 the directory restored without its\nsidecars, or hand-pruned \u2014 the 3MF still carries the same provenance, so the\nlink keeps working from the file alone.",
        "operationId": "get_edit_target_api_v1_outputs__output_id__edit_get",
        "parameters": [
          {
            "in": "path",
            "name": "output_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Output Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditTarget"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve an edit deep link",
        "tags": [
          "outputs"
        ]
      }
    },
    "/api/v1/outputs/{output_id}/geometry": {
      "get": {
        "description": "Printability measurements of the output's closed per-colour solids (#284):\nopen and non-manifold edges with their locations, bounding box, bed contact,\nheight-to-base ratio, overhang area by angle, and estimates of the thinnest\nwall and smallest feature. Coordinates are the model's own (mm, Z up), as in\nthe preview. Computed on first ask and cached beside the output.",
        "operationId": "get_output_geometry_api_v1_outputs__output_id__geometry_get",
        "parameters": [
          {
            "in": "path",
            "name": "output_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Output Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeometryAnalysis"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mesh geometry analysis",
        "tags": [
          "outputs"
        ]
      }
    },
    "/api/v1/outputs/{output_id}/model.3mf": {
      "get": {
        "operationId": "download_output_api_v1_outputs__output_id__model_3mf_get",
        "parameters": [
          {
            "in": "path",
            "name": "output_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Output Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "model/3mf": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download the 3MF",
        "tags": [
          "outputs"
        ]
      }
    },
    "/api/v1/outputs/{output_id}/plates": {
      "get": {
        "description": "What the print picker offers as ``plate_id`` (#83). ScadBuddy's own renders are\nalways one plate, which the picker does not ask about.",
        "operationId": "get_output_plates_api_v1_outputs__output_id__plates_get",
        "parameters": [
          {
            "in": "path",
            "name": "output_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Output Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OutputPlate"
                  },
                  "title": "Response Get Output Plates Api V1 Outputs  Output Id  Plates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The 3MF's plates",
        "tags": [
          "outputs"
        ]
      }
    },
    "/api/v1/outputs/{output_id}/plates/{index}/thumbnail": {
      "get": {
        "description": "The plate's own cover from inside the 3MF, the one Bambu Studio would show.",
        "operationId": "get_output_plate_thumbnail_api_v1_outputs__output_id__plates__index__thumbnail_get",
        "parameters": [
          {
            "in": "path",
            "name": "output_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Output Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "index",
            "required": true,
            "schema": {
              "title": "Index",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/png": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A plate's cover image",
        "tags": [
          "outputs"
        ]
      }
    },
    "/api/v1/outputs/{output_id}/send": {
      "post": {
        "description": "Upload ``model.3mf`` to the configured library folder and, in ``queue`` mode,\nslice and queue it.\n\nThe file is read from the PVC and pushed by the server, so the API key never\nreaches the browser. A re-send replaces the file Bambuddy already holds rather\nthan adding a second copy.",
        "operationId": "send_output_to_bambuddy_api_v1_outputs__output_id__send_post",
        "parameters": [
          {
            "in": "path",
            "name": "output_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Output Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send the 3MF to Bambuddy",
        "tags": [
          "outputs"
        ]
      }
    },
    "/api/v1/outputs/{output_id}/thumbnail": {
      "get": {
        "operationId": "get_output_thumbnail_api_v1_outputs__output_id__thumbnail_get",
        "parameters": [
          {
            "in": "path",
            "name": "output_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Output Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/png": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Output thumbnail",
        "tags": [
          "outputs"
        ]
      },
      "put": {
        "operationId": "put_output_thumbnail_api_v1_outputs__output_id__thumbnail_put",
        "parameters": [
          {
            "in": "path",
            "name": "output_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}$",
              "title": "Output Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_put_output_thumbnail_api_v1_outputs__output_id__thumbnail_put"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload the canvas capture",
        "tags": [
          "outputs"
        ]
      }
    },
...
  }
}